datagen/pyproject.toml
#
# boxes - Command line filter to draw/remove ASCII boxes around text
# Copyright (c) 1999-2024 Thomas Jensen and the boxes contributors
#
# This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public
# License, version 3, as published by the Free Software Foundation.
# This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied
# warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
# details.
# You should have received a copy of the GNU General Public License along with this program.
# If not, see <https://www.gnu.org/licenses/>.
#

[project]
name = "datagen"
description = "Convert boxes config to Jekyll data file for the website's design explorer"
version = "0.1.0"
authors = [ { name = "Thomas Jensen", email = "boxes@thomasjensen.com"} ]
maintainers = [ { name = "Thomas Jensen", email = "boxes@thomasjensen.com" } ]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
requires-python = ">= 3.10"
dependencies = [
    "ansi2html ==1.9.1",
    "Click ==8.1.7",
    "PyYAML ==6.0.1"
]

[project.scripts]
convert = "datagen.main:main"


[build-system]
requires = [
    "setuptools >= 69.0.3"
]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["datagen"]
